+++ wishlists/serializers.py
@@ -1,10 +1,10 @@
 from rest_framework.serializers import ModelSerializer
+
 from rooms.serializers import RoomListSerializer
-from .models import Wishlist
+from wishlists.models.wish_list import Wishlist
 
 
 class WishlistSerializer(ModelSerializer):
-
     rooms = RoomListSerializer(
         many=True,
         read_only=True,
