--- wishlists/views.py
@@ -1,102 +1,0 @@
-from rest_framework.exceptions import NotFound
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.status import HTTP_200_OK
-from rest_framework.views import APIView
-
-from rooms.models.room import Room
-
-from .models import Wishlist
-from .serializers import WishlistSerializer
-
-
-class Wishlists(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        all_wishlists = Wishlist.objects.filter(user=request.user)
-        serializer = WishlistSerializer(
-            all_wishlists,
-            many=True,
-            context={"request": request},
-        )
-        return Response(serializer.data)
-
-    def post(self, request):
-        serializer = WishlistSerializer(data=request.data)
-        if serializer.is_valid():
-            wishlist = serializer.save(
-                user=request.user,
-            )
-            serializer = WishlistSerializer(wishlist)
-            return Response(serializer.data)
-        return Response(serializer.errors)
-
-
-class WishlistDetail(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get_object(self, pk, user):
-        try:
-            return Wishlist.objects.get(pk=pk, user=user)
-        except Wishlist.DoesNotExist:
-            raise NotFound
-
-    def get(self, request, pk):
-        wishlist = self.get_object(pk, request.user)
-        serializer = WishlistSerializer(
-            wishlist,
-            context={"request": request},
-        )
-        return Response(serializer.data)
-
-    def delete(self, request, pk):
-        wishlist = self.get_object(pk, request.user)
-        wishlist.delete()
-        return Response(status=HTTP_200_OK)
-
-    def put(self, request, pk):
-        wishlist = self.get_object(pk, request.user)
-        serializer = WishlistSerializer(
-            wishlist,
-            data=request.data,
-            partial=True,
-        )
-        if serializer.is_valid():
-            wishlist = serializer.save()
-            serializer = WishlistSerializer(
-                wishlist,
-                context={"request": request},
-            )
-            return Response(serializer.data)
-
-        return Response(serializer.errors)
-
-
-class WishlistToggle(APIView):
-    def get_list(self, pk, user):
-        try:
-            return Wishlist.objects.get(pk=pk, user=user)
-        except Wishlist.DoesNotExist:
-            raise NotFound
-
-    def get_room(self, pk):
-        try:
-            return Room.objects.get(pk=pk)
-        except Room.DoesNotExist:
-            raise NotFound
-
-    def put(self, request, room_pk):
-        wishlist = Wishlist.objects.filter(user=request.user).first()
-
-        if wishlist is None:
-            wishlist = Wishlist.objects.create(
-                user=request.user, name=f"{request.user.name} 님의 찜 목록"
-            )
-
-        room = self.get_room(room_pk)
-        if wishlist.rooms.filter(pk=room_pk).exists():
-            wishlist.rooms.remove(room)
-        else:
-            wishlist.rooms.add(room)
-        return Response(status=HTTP_200_OK)
